Extract archive execution DAO from archvie configuration class
To prevent circular dependency between ArchivedIndexDao and
PostgresArchiveDAOConfiguration. Now that ArchivedIndexDao depends on
execution DAO, there is a loop between ArchivedIndexDao, ExecutionDao
and PostgresArchiveDao.
